Refactor. location path and folder name
