+++ speech-assistant-proxy/default.conf
@@ -13,14 +13,9 @@ server {
         try_files $uri $uri/ =404;
     }
 
-    location /speech-assistant/update/feature {
+    location /speech-assistant/feature {
         default_type text/plain;
         js_content featureToggle.update;
-    }
-
-    location /speech-assistant/get/feature {
-        default_type text/plain;
-        js_content featureToggle.get;
     }
     
 }
